moves notification api to a separate modules
