+++ apps/judicial-system/backend/src/app/modules/case/dto/index.ts
@@ -1,4 +1,3 @@
 export { CreateCaseDto } from './createCase.dto'
 export { UpdateCaseDto } from './updateCase.dto'
 export { TransitionCaseDto } from './transitionCase.dto'
-export { SendNotificationDto } from './sendNotification.dto'
