+++ apps/judicial-system/backend/src/app/modules/case/index.ts
@@ -1,1 +1,2 @@
 export { CaseModule } from './case.module'
+export { CaseService } from './case.service'
