+++ apps/judicial-system/backend/src/app/modules/case/models/case.model.ts
@@ -21,7 +21,7 @@ import {
 } from '@island.is/judicial-system/types'
 
 import { User } from '../../user'
-import { Notification } from './notification.model'
+import { Notification } from '../../notification/notification.model'
 
 @Table({
   tableName: 'case',
