+++ apps/judicial-system/backend/src/app/modules/case/models/index.ts
@@ -1,3 +1,2 @@
 export { Case } from './case.model'
-export { Notification } from './notification.model'
 export { SignatureResponse } from './signatureResponse.model'
