+++ apps/judicial-system/backend/src/app/modules/notification/notification.model.ts
@@ -11,7 +11,7 @@ import { ApiProperty } from '@nestjs/swagger'
 
 import { NotificationType } from '@island.is/judicial-system/types'
 
-import { Case } from './case.model'
+import { Case } from '../case/models/case.model'
 
 @Table({
   tableName: 'notification',
